Перенес showError() в utils
