+++ js/utils.js
@@ -1,3 +1,5 @@
+const MESSAGE_TIME = 4000;
+
 function getRandomFromRange (min, max) {
   if (Number.isInteger(min) && Number.isInteger(max)) {
     if (min >= 0 && max >= 0 && min <= max) {
@@ -26,4 +28,25 @@ function hasDuplicates (array) {
   return false;
 }
 
-export {getRandomFromRange, getRandomArrayElement, isEscEvent, hasDuplicates};
+function showError (message) {
+  const alertContainer = document.createElement('div');
+  alertContainer.style.zIndex = 100;
+  alertContainer.style.position = 'absolute';
+  alertContainer.style.left = 0;
+  alertContainer.style.top = 0;
+  alertContainer.style.right = 0;
+  alertContainer.style.padding = '10px 3px';
+  alertContainer.style.fontSize = '20px';
+  alertContainer.style.textAlign = 'center';
+  alertContainer.style.backgroundColor = 'red';
+
+  alertContainer.textContent = message;
+
+  document.body.append(alertContainer);
+
+  setTimeout(() => {
+    alertContainer.remove();
+  }, MESSAGE_TIME);
+}
+
+export {getRandomFromRange, getRandomArrayElement, isEscEvent, hasDuplicates, showError};
